Other minor revisions
While this branch was created to allow the master process to show the
termination of the map process this cannot be done because map is not directly a
child of the master process but the server one. Server process could be modified
to show the termination of the process but no good solution is available infact
the possibile ones are that in the server a fork is created and one of the
childs waits for the termination of map, or a nonblocking wait. The problem is
that in the server there is a 2 seconds sleep making this solution unreliable
and unresponsive.